+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## 0.0.29
+
+- chore: add nigeria utils
+
+## 0.0.28
+
+- chore: Edit information dropdown widget
 
 ## 0.0.27